Handle unknown fault codes without raising exception
