+++ OmniKit/MessageTransport/MessageBlocks/PodInfoDataLog.swift
@@ -15,13 +15,13 @@ public struct PodInfoDataLog : PodInfo {
     // public let numberOfWords: UInt8 = 60
     // public let numberOfBytes: UInt8 = 10
     
-    public var podInfoType              : PodInfoResponseSubType = .dataLog
-    public let faultEventType           : PodInfoResponseSubType.FaultEventType
-    public let data                     : Data
-    public let timeFaultEvent           : TimeInterval
-    public let timeActivation           : TimeInterval
-    public let dataChunkSize            : UInt8
-    public let dataChunkWords           : UInt8
+    public var podInfoType: PodInfoResponseSubType = .dataLog
+    public let faultEventCode: FaultEventCode
+    public let data: Data
+    public let timeFaultEvent: TimeInterval
+    public let timeActivation: TimeInterval
+    public let dataChunkSize: UInt8
+    public let dataChunkWords: UInt8
     // TODO adding a datadump variable based on length
     //length = dataChunckSize * N + 8 where N is the number of dword entries
     // public let loggedData               : Data
@@ -30,12 +30,12 @@ public struct PodInfoDataLog : PodInfo {
         if encodedData.count < Int(6) {
             throw MessageBlockError.notEnoughData
         }
-        self.podInfoType         = PodInfoResponseSubType(rawValue: encodedData[0])!
-        self.faultEventType      = PodInfoResponseSubType.FaultEventType(rawValue: encodedData[1])!
-        self.timeFaultEvent      = TimeInterval(minutes: Double((Int(encodedData[2] & 0b1) << 8) + Int(encodedData[3])))
-        self.timeActivation      = TimeInterval(minutes: Double((Int(encodedData[4] & 0b1) << 8) + Int(encodedData[5])))
-        self.dataChunkSize       = encodedData[6]
-        self.dataChunkWords      = encodedData[7]
+        self.podInfoType = PodInfoResponseSubType(rawValue: encodedData[0])!
+        self.faultEventCode = FaultEventCode(rawValue: encodedData[1])
+        self.timeFaultEvent = TimeInterval(minutes: Double((Int(encodedData[2] & 0b1) << 8) + Int(encodedData[3])))
+        self.timeActivation = TimeInterval(minutes: Double((Int(encodedData[4] & 0b1) << 8) + Int(encodedData[5])))
+        self.dataChunkSize = encodedData[6]
+        self.dataChunkWords = encodedData[7]
         
         // self.loggedData          = encodedData[10...encodedData.count]
         // self.dataFromFlashMemory = Data(encodedData[22...124])
