+++ OmniKit/MessageTransport/MessageBlocks/PodInfoFaultEvent.swift
@@ -26,21 +26,21 @@ public struct PodInfoFaultEvent : PodInfo {
     // public let numberOfBytes: UInt8 = 10
     
     public var podInfoType: PodInfoResponseSubType = .faultEvents
-    public let reservoirStatus: StatusResponse.ReservoirStatus
+    public let reservoirStatus: ReservoirStatus
     public let deliveryType: DeliveryType
     public let insulinNotDelivered: Double
     public let podMessageCounter: UInt8
     public let unknownPageCode: Double
-    public let originalLoggedFaultEvent: PodInfoResponseSubType.FaultEventType
+    public let originalLoggedFaultEvent: FaultEventCode
     public let faultEventTimeSinceActivation: Double
     public let insulinRemaining: Double
     public let timeActive: TimeInterval
-    public let secondaryLoggedFaultEvent: PodInfoResponseSubType.FaultEventType
+    public let secondaryLoggedFaultEvent: FaultEventCode
     public let logEventError: Bool
-    public let reservoirStatusAtFirstLoggedFaultEvent: StatusResponse.ReservoirStatus
+    public let reservoirStatusAtFirstLoggedFaultEvent: ReservoirStatus
     public let recieverLowGain: UInt8
     public let radioRSSI: UInt8
-    public let reservoirStatusAtFirstLoggedFaultEventCheck: StatusResponse.ReservoirStatus
+    public let reservoirStatusAtFirstLoggedFaultEventCheck: ReservoirStatus
     public let insulinStateTableCorruption: Bool
     public let immediateBolusInProgress: Bool
     
@@ -52,8 +52,8 @@ public struct PodInfoFaultEvent : PodInfo {
             throw MessageBlockError.notEnoughData
         }
         
-        guard let reservoirStatus = StatusResponse.ReservoirStatus(rawValue: encodedData[1]) else {
-            throw MessageError.unknownValue(value: encodedData[1], typeDescription: "StatusResponse.ReservoirStatus")
+        guard let reservoirStatus = ReservoirStatus(rawValue: encodedData[1]) else {
+            throw MessageError.unknownValue(value: encodedData[1], typeDescription: "ReservoirStatus")
         }
         self.reservoirStatus = reservoirStatus
         
@@ -64,10 +64,7 @@ public struct PodInfoFaultEvent : PodInfo {
         self.podMessageCounter = encodedData[5]
         self.unknownPageCode = Double(Int(encodedData[6]) | Int(encodedData[7]))
         
-        guard let originalLoggedFaultEvent = PodInfoResponseSubType.FaultEventType(rawValue: encodedData[8]) else {
-            throw MessageError.unknownValue(value: encodedData[8], typeDescription: "FaultEventType")
-        }
-        self.originalLoggedFaultEvent = originalLoggedFaultEvent
+        self.originalLoggedFaultEvent = FaultEventCode(rawValue: encodedData[8])
         
         self.faultEventTimeSinceActivation = TimeInterval(minutes: Double((Int(encodedData[9] & 0b1) << 8) + Int(encodedData[10])))
         
@@ -75,14 +72,14 @@ public struct PodInfoFaultEvent : PodInfo {
         
         self.timeActive = TimeInterval(minutes: Double((Int(encodedData[13] & 0b1) << 8) + Int(encodedData[14])))
         
-        self.secondaryLoggedFaultEvent = PodInfoResponseSubType.FaultEventType(rawValue: encodedData[15])!
+        self.secondaryLoggedFaultEvent = FaultEventCode(rawValue: encodedData[15])
         
         self.logEventError = encodedData[16] == 2
 
         self.insulinStateTableCorruption = encodedData[17] & 0b10000000 != 0
         
-        guard let reservoirStatusAtFirstLoggedFaultEventType = StatusResponse.ReservoirStatus(rawValue: encodedData[17] & 0xF) else {
-            throw MessageError.unknownValue(value: encodedData[17] & 0xF, typeDescription: "ProgressType")
+        guard let reservoirStatusAtFirstLoggedFaultEventType = ReservoirStatus(rawValue: encodedData[17] & 0xF) else {
+            throw MessageError.unknownValue(value: encodedData[17] & 0xF, typeDescription: "ReservoirStatus")
         }
         
         self.immediateBolusInProgress = encodedData[17] & 0b00010000 != 0
@@ -93,8 +90,8 @@ public struct PodInfoFaultEvent : PodInfo {
         
         self.radioRSSI =  encodedData[18] & 0xF
         
-        guard let reservoirStatusAtFirstLoggedFaultEventCheckType = StatusResponse.ReservoirStatus(rawValue: encodedData[19] & 0xF) else {
-            throw MessageError.unknownValue(value: encodedData[19] & 0xF, typeDescription: "ProgressType")
+        guard let reservoirStatusAtFirstLoggedFaultEventCheckType = ReservoirStatus(rawValue: encodedData[19] & 0xF) else {
+            throw MessageError.unknownValue(value: encodedData[19] & 0xF, typeDescription: "ReservoirStatus")
         }
         self.reservoirStatusAtFirstLoggedFaultEventCheck = reservoirStatusAtFirstLoggedFaultEventCheckType
